requirements.txt
langchain_ollama==1.0.0
pypdf==6.1.3
Requests==2.32.5
tqdm==4.64.1
weaviate_client==4.17.0
